# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

const maxListingYear = new Date().getFullYear() + 1;

const OptionalTrimmedString = zod.preprocess(
  (value) => (typeof value === "string" && value.trim() === "" ? undefined : value),
  zod.string().trim().min(1).optional(),
);

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

export const ListingSchema = zod.object({
  id: zod.number(),
  title: zod.string(),
  marque: zod.string(),
  modele: zod.string(),
  year: zod.number(),
  kmRaw: zod.number(),
  km: zod.string(),
  fuel: zod.enum(["Essence", "Diesel", "GPL", "Électrique", "Hybride"]),
  transmission: zod.enum(["Manuelle", "Automatique"]),
  location: zod.string(),
  wilaya: zod.string(),
  priceRaw: zod.number(),
  price: zod.string(),
  color: zod.string(),
  verified: zod.boolean(),
  badge: zod.string().optional(),
});

export const ListingDetailSchema = zod.object({
  id: zod.number(),
  description: zod.string(),
  sellerName: zod.string(),
  sellerType: zod.enum(["Particulier", "Concessionnaire"]),
  sellerMember: zod.string(),
  sellerPhone: zod.string(),
  sellerWhatsapp: zod.string(),
  sellerRating: zod.number(),
  sellerTotalAds: zod.number(),
  options: zod.array(zod.string()),
  couleur: zod.string(),
  portes: zod.number(),
  places: zod.number(),
  puissance: zod.number(),
  cylindree: zod.string(),
  etat: zod.enum(["Excellent", "Très bon", "Bon", "Passable"]),
  firstHand: zod.boolean(),
  dedouane: zod.boolean(),
  views: zod.number(),
  postedDaysAgo: zod.number(),
});

export const ListingFacetsSchema = zod.object({
  marques: zod.array(zod.string()),
  wilayas: zod.array(zod.string()),
  fuels: zod.array(ListingSchema.shape.fuel),
  transmissions: zod.array(ListingSchema.shape.transmission),
});

export const ListingsQuerySchema = zod.object({
  search: zod.string().optional(),
  marque: zod.string().optional(),
  wilaya: zod.string().optional(),
  fuels: zod.string().optional(),
  transmissions: zod.string().optional(),
  priceMin: zod.coerce.number().optional(),
  priceMax: zod.coerce.number().optional(),
  yearMin: zod.coerce.number().optional(),
  yearMax: zod.coerce.number().optional(),
  kmMin: zod.coerce.number().optional(),
  kmMax: zod.coerce.number().optional(),
  verifiedOnly: zod
    .preprocess(
      (value) =>
        value === "true" ? true : value === "false" ? false : value,
      zod.boolean(),
    )
    .optional(),
  sort: zod.enum(["recent", "prix_asc", "prix_desc", "km_asc", "annee_desc"]).optional(),
  page: zod.coerce.number().int().positive().optional(),
  pageSize: zod.coerce.number().int().positive().max(50).optional(),
});

export const ListListingsResponse = zod.object({
  items: zod.array(ListingSchema),
  total: zod.number(),
  page: zod.number(),
  pageSize: zod.number(),
  totalPages: zod.number(),
  facets: ListingFacetsSchema,
});

export const ListingDetailResponse = zod.object({
  listing: ListingSchema,
  detail: ListingDetailSchema,
  similar: zod.array(ListingSchema),
});

export const CreateListingPhotoSchema = zod.object({
  url: zod.string().url(),
  alt: OptionalTrimmedString,
  position: zod.coerce.number().int().min(0).optional(),
  isPrimary: zod.boolean().optional(),
});

export const CreateListingRequest = zod.object({
  vehicleType: OptionalTrimmedString,
  marque: zod.string().trim().min(1).max(80),
  modele: zod.string().trim().min(1).max(80),
  year: zod.coerce.number().int().min(1950).max(maxListingYear),
  kmRaw: zod.coerce.number().int().min(0).max(2_000_000),
  fuel: ListingSchema.shape.fuel,
  transmission: ListingSchema.shape.transmission,
  location: zod.string().trim().min(1).max(160),
  wilaya: zod.string().trim().min(1).max(80),
  priceRaw: zod.coerce.number().int().positive().max(1_000_000_000),
  color: zod.string().trim().min(1).max(120).optional(),
  description: zod.string().trim().max(1000).optional(),
  couleur: OptionalTrimmedString,
  portes: zod.coerce.number().int().min(2).max(6).optional(),
  places: zod.coerce.number().int().min(1).max(9).optional(),
  puissance: zod.coerce.number().int().min(0).max(3000).optional(),
  cylindree: OptionalTrimmedString,
  condition: ListingDetailSchema.shape.etat.optional(),
  firstHand: zod.boolean().optional(),
  dedouane: zod.boolean().optional(),
  options: zod.array(zod.string().trim().min(1).max(80)).max(40).optional(),
  photos: zod.array(CreateListingPhotoSchema).max(10).optional(),
  seller: zod.object({
    name: zod.string().trim().min(2).max(160),
    email: OptionalTrimmedString.pipe(zod.string().email().optional()),
    phone: zod.string().trim().min(6).max(40),
    whatsapp: OptionalTrimmedString,
    wilaya: zod.string().trim().min(1).max(80),
    sellerType: zod.enum(["particulier", "concessionnaire"]).default("particulier"),
  }),
});

export const CreateListingResponse = ListingDetailResponse;